Created copy of Remi to make AM Adaptive.
