+++ pints/_mcmc/_adaptive_covariance_am_global_scale.py
